File uploading and downloading

+++ src/main/java/com/example/javaserver/repo/UserRepo.java
@@ -3,7 +3,7 @@ package com.example.javaserver.repo;
 import com.example.javaserver.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<User,Long> {
+public interface UserRepo extends JpaRepository<User,Integer> {
     boolean existsByLogin(String login);
     User getUserByLogin(String login);
 }
